+++ os_remote/client/os_client_wrap.cc
@@ -14,15 +14,15 @@ typedef unsigned int u32;
 std::unique_ptr <os_remote::OSRemote::Stub> stub_ = os_remote::OSRemote::NewStub(
         grpc::CreateChannel("127.0.0.1:50051", grpc::InsecureChannelCredentials()));
 
-extern "C" const char *clientInit() {
-    os_remote::InitRequest request;
-    os_remote::InitReply reply;
+extern "C" const char *clientInit(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
     ClientContext context;
 
     Status status = stub_->Init(&context, request, &reply);
 
     if (status.ok()) {
-        return reply.rc().data();
+        return reply.outarg().data();
     } else {
         std::cout << status.error_code() << ": " << status.error_message() << std::endl;
         return "RPC failed";
@@ -220,6 +220,184 @@ extern "C" const char *clientRead(char *argin, u32 inlen, u32 outlen) {
     request.set_outlen(outlen);
 
     Status status = stub_->Read(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientTruncate(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->Truncate(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientSync(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->Sync(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientFileSize(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->FileSize(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientFileControl(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->FileControl(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientSectorSize(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->SectorSize(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientDeviceCharacteristics(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->DeviceCharacteristics(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientClose(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->Close(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+extern "C" const char *clientLock(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->Lock(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+extern "C" const char *clientUnlock(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->Unlock(&context, request, &reply);
+
+    if (status.ok()) {
+        return reply.outarg().data();
+    } else {
+        std::cout << status.error_code() << ": " << status.error_message() << std::endl;
+        return "RPC failed";
+    }
+}
+
+extern "C" const char *clientCheckReservedLock(char *argin, u32 inlen, u32 outlen) {
+    os_remote::ArgRequest request;
+    os_remote::ArgReply reply;
+    ClientContext context;
+
+    request.set_inarg(argin, inlen);
+    request.set_outlen(outlen);
+
+    Status status = stub_->CheckReservedLock(&context, request, &reply);
 
     if (status.ok()) {
         return reply.outarg().data();
