+++ os_remote/client/util.c
