+++ os_remote/server/os_server_wrap.c
@@ -30,20 +30,23 @@ static sqlite3_vfs aVfs[] = {
         UNIXVFS("unix", posixIoFinder),
 };
 
-char *WrapInit(void) {
+void WrapInit(const char *argIn, char *argOut){
+    printf("---WrapInit:\n");
     int rc = sqlite3_os_init();
-    char *p = (char *) (&rc);
-    return p;
+    sqlite3_os_initConvertReturnToChar(&rc, argOut);
 }
 
 void WrapOpen(const char *argIn, char *argOut) {
+    printf("---WrapOpen:\n");
     char path[512];
     unixFile file_infor;
     int in_flags;
     int out_flags;
 
     unixOpenConvertCharToArgIn(argIn, &aVfs[0], path, (sqlite3_file *) &file_infor, &in_flags, &out_flags);
+    printf("   id.h = %d\n", file_infor.h);
     int rc = unixOpen(&aVfs[0], path, (sqlite3_file *) &file_infor, in_flags, &out_flags);
+    printf("   id.h = %d\n", file_infor.h);
     unixOpenConvertReturnToChar((sqlite3_file *) &file_infor, &out_flags, &rc, argOut);
 }
 
@@ -128,7 +131,7 @@ void WrapWrite(const char *argIn, char *argOut) {
     sqlite3_int64 offset;
     unixWriteConvertCharToArgIn(argIn, (sqlite3_file *) &id, pBuf, &amt, &offset);
     int rc = unixWrite((sqlite3_file *) &id, pBuf, amt, offset);
-    unixWriteConvertReturnToChar((sqlite3_file *) &id, pBuf,&amt, &rc, argOut);
+    unixWriteConvertReturnToChar((sqlite3_file *) &id, pBuf, &amt, &rc, argOut);
 }
 
 void WrapRead(const char *argIn, char *argOut) {
@@ -140,4 +143,95 @@ void WrapRead(const char *argIn, char *argOut) {
     unixReadConvertCharToArgIn(argIn, (sqlite3_file *) &id, pBuf, &amt, &offset);
     int rc = unixRead((sqlite3_file *) &id, pBuf, amt, offset);
     unixReadConvertReturnToChar((sqlite3_file *) &id, pBuf, &amt, &rc, argOut);
+}
+
+void WrapTruncate(const char *argIn, char *argOut) {
+    unixFile id;
+    i64 nByte;
+
+    unixTruncateConvertCharToArgIn(argIn, (sqlite3_file *) &id, &nByte);
+    int rc = unixTruncate((sqlite3_file *) &id, nByte);
+    unixTruncateConvertReturnToChar((sqlite3_file *) &id, &rc, argOut);
+}
+void WrapSync(const char *argIn, char *argOut) {
+    unixFile id;
+    int flags;
+
+    unixSyncConvertCharToArgIn(argIn, (sqlite3_file *) &id, &flags);
+    int rc = unixSync((sqlite3_file *) &id, flags);
+    unixSyncConvertReturnToChar((sqlite3_file *) &id, &rc, argOut);
+}
+void WrapFileSize(const char *argIn, char *argOut) {
+    unixFile id;
+    i64 pSize;
+
+    unixFileSizeConvertCharToArgIn(argIn, (sqlite3_file *) &id, &pSize);
+    int rc = unixFileSize((sqlite3_file *) &id, &pSize);
+    unixFileSizeConvertReturnToChar((sqlite3_file *) &id, &pSize, &rc, argOut);
+}
+void WrapFileControl(const char *argIn, char *argOut) {
+    unixFile id;
+    int op;
+    int pArg;
+
+    unixFileControlConvertCharToArgIn(argIn, (sqlite3_file *) &id, &op, &pArg);
+    int rc = unixFileControl((sqlite3_file *) &id, op, &pArg);
+    unixFileControlConvertReturnToChar((sqlite3_file *) &id, &pArg, &rc, argOut);
+}
+void WrapSectorSize(const char *argIn, char *argOut) {
+    unixFile id;
+
+    unixSectorSizeConvertCharToArgIn(argIn, (sqlite3_file *) &id);
+    int sectorSize = unixSectorSize((sqlite3_file *) &id);
+    unixSectorSizeConvertReturnToChar((sqlite3_file *) &id, &sectorSize, argOut);
+}
+void WrapDeviceCharacteristics(const char *argIn, char *argOut) {
+    unixFile id;
+
+    unixDeviceCharacteristicsConvertCharToArgIn(argIn, (sqlite3_file *) &id);
+    int deviceCharacteristics = unixDeviceCharacteristics((sqlite3_file *) &id);
+    unixDeviceCharacteristicsConvertReturnToChar((sqlite3_file *) &id, &deviceCharacteristics, argOut);
+}
+void WrapClose(const char *argIn, char *argOut) {
+    printf("---WrapClose:\n");
+    unixFile id;
+
+    unixCloseConvertCharToArgIn(argIn, (sqlite3_file *) &id);
+    printf("   id.h = %d\n", id.h);
+    int rc = unixClose((sqlite3_file *) &id);   // TODO: 上面一个能输出，下面一个输出不了，出现了段错误
+    printf("   id.h = %d\n", id.h);
+    unixCloseConvertReturnToChar((sqlite3_file *) &id, &rc, argOut);
+}
+void WrapLock(const char *argIn, char *argOut) {
+    printf("---WrapLock:\n");
+    unixFile id;
+    int eFileLock;
+
+    unixLockConvertCharToArgIn(argIn, (sqlite3_file *) &id, &eFileLock);
+    printf("   id.h = %d\n", id.h);
+    int rc = unixLock((sqlite3_file *) &id, eFileLock);
+    printf("   id.h = %d\n", id.h);
+    unixLockConvertReturnToChar((sqlite3_file *) &id, &rc, argOut);
+}
+void WrapUnlock(const char *argIn, char *argOut) {
+    printf("---WrapUnlock:\n");
+    unixFile id;
+    int eFileLock;
+
+    unixUnlockConvertCharToArgIn(argIn, (sqlite3_file *) &id, &eFileLock);
+    printf("   id.h = %d\n", id.h);
+    int rc = unixUnlock((sqlite3_file *) &id, eFileLock);
+    printf("   id.h = %d\n", id.h);
+    unixUnlockConvertReturnToChar((sqlite3_file *) &id, &rc, argOut);
+}
+void WrapCheckReservedLock(const char *argIn, char *argOut) {
+    printf("---WrapCheckReservedLock:\n");
+    unixFile id;
+    int pResOut = 0;
+
+    unixCheckReservedLockConvertCharToArgIn(argIn, (sqlite3_file *) &id, &pResOut);
+    printf("   id.h = %d\n", id.h);
+    int rc = unixCheckReservedLock((sqlite3_file *) &id, &pResOut);
+    printf("   id.h = %d\n", id.h);
+    unixCheckReservedLockConvertReturnToChar((sqlite3_file *) &id, &pResOut, &rc, argOut);
 }
